requirements.txt
python-slugify
ics
requests
pyyaml